+++ CHANGELOG.md
@@ -5,6 +5,8 @@ which adhereth unto [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+## [v2.2.0] (2021-03-09)
+
 ### Added
 
 - Optional attribute specification for inline-style images
@@ -54,7 +56,8 @@ which adhereth unto [Semantic Versioning](https://semver.org/).
 
 - Initial release.
 
-[Unreleased]: https://github.com/conway-markdown/conway-markdown/compare/v2.1.0...HEAD
+[Unreleased]: https://github.com/conway-markdown/conway-markdown/compare/v2.2.0...HEAD
+[v2.2.0]: https://github.com/conway-markdown/conway-markdown/compare/v2.1.0...v2.2.0
 [v2.1.0]: https://github.com/conway-markdown/conway-markdown/compare/v2.0.1...v2.1.0
 [v2.0.1]: https://github.com/conway-markdown/conway-markdown/compare/v2.0.0...v2.0.1
 [v2.0.0]: https://github.com/conway-markdown/conway-markdown/compare/v1.0.0...v2.0.0